Consolidate utils files and update endpoint
