+++ src/components/user-reservation/reservation/TimeSlotList.jsx
@@ -19,7 +19,7 @@ const TimeSlotList = () => {
   }, [reservationsList, hourRange, selectedGarage.spots]);
 
   return (
-    <details className="dropdown max-w-[500px] w-48">
+    <details className="dropdown max-w-[300px] w-48">
       <summary className="btn w-full">
         <svg
           xmlns="http://www.w3.org/2000/svg"
